+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 1.6.0 - 2020-07-23
+
+### Added
+- Added new property `maxAttempts` to the `LineItem` entity.
+- Added support for ingestion of `maxAttempts` on the `LineItemIngester`.
+- Added new property `attemptsCount` to the `Assignment` entity.
+
+### Changed
+- Changed LTI outcome state update to `ready` if `Assignment` has additional attempts available.
+- Changed garbage collection state update to `ready` if `Assignment` has additional attempts available.
+- Increment the `attemptsCount` on upon LTI Launch if state is not started.
+- Set default value of `0` for `attemptsCount` for new assignments during user ingestion.
+- The `/api/v1/assignments` endpoint now returns all users assignments (available or not).
+- The `UserCacheInvalidationSubscriber` now warms up the cache after invalidating it.
+
 ## 1.5.0 - 2020-06-17
 
 ### Added